Update README.md
Understanding how government expenditure in agriculture affects employment opportunities to aid poverty reduction, with the expectation of likely associated challenges in Nigeria, remains the core objective of this work in investigating the effects of government agricultural spending, unemployment, and poverty reduction in Nigeria (1981 – 2020). Based on co-integrations and causality effects reported by the Vector Error Correction Model, continuous increases of agricultural output, Gross Domestic Product (as income), Government expenditure in agriculture, Investment, and the Rate of education affect employment opportunities in agriculture inversely. Likewise, it has been empirically established through years of investigation, relevant available data, and variables that Gross Domestic Saving is desirable for better lives in the face of a high Rate of poverty in Nigeria; the agricultural sector becomes an employment opportunity for many. One of the empirical principles/policies to alleviate poverty in Nigeria is to first reduce it by providing employment opportunities in agriculture backed up by quality education as a guarantor for ever-increasing agricultural output. Therefore, this work supports government agricultural spending on employment opportunities to reduce poverty.